% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tactileR-package.R
\docType{package}
\name{tactileR-package}
\alias{tactileR}
\alias{tactileR-package}
\title{tactileR: Converting R Graphics into a Braille Ready-to-Print PDF}
\description{
Supporting for blind or visually impaired R users who have an access to Swell Form Machine or braille embosser to print R graphics in a tactile form with Braille font embedded.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jooyoungseo/tactileR}
  \item Report bugs at \url{https://github.com/jooyoungseo/tactileR/issues}
}

}
\author{
\strong{Maintainer}: JooYoung Seo \email{jooyoung@psu.edu}

Other contributors:
\itemize{
  \item Tsan-Kuang Lee \email{developer@tklee.com} [contributor]
}

}
\keyword{internal}
